Create new, simplified gallery route for /explore
